refactored with mult utilities
