+++ src/plugins/blocksmith/utils/getLines.js
@@ -1,0 +1,48 @@
+import DraftOffsetKey from 'draft-js/lib/DraftOffsetKey';
+
+export default (editorState, testClass) => {
+  let moreThanOneLine = false;
+  let height;
+  let testSpanHeight;
+  let line = '';
+  let startPoint = 0;
+  const endWordIndices = [];
+  const lines = [];
+  const anchorKey = editorState.getSelection().getAnchorKey();
+  const offsetKey = DraftOffsetKey.encode(anchorKey, 0, 0);
+  const currentBlockNode = document.querySelectorAll(`[data-offset-key="${offsetKey}"]`)[0];
+  const computedStyle = getComputedStyle(currentBlockNode);
+  const width = +computedStyle.width.replace('px', '');
+  const testSpan = document.createElement('span');
+  testSpan.classList.add(testClass);
+  testSpan.setAttribute('style', `width: ${width}px`);
+  document.body.appendChild(testSpan);
+  const text = editorState.getCurrentContent().getBlockForKey(anchorKey).getText();
+  const words = text.split(' ');
+  testSpan.innerHTML = words[0];
+  height = testSpan.getBoundingClientRect().height;
+  for (let i = 1; i < words.length; i += 1) {
+    testSpan.innerHTML += ` ${words[i]}`;
+    testSpanHeight = testSpan.getBoundingClientRect().height;
+    if (testSpanHeight > height) {
+      moreThanOneLine = true;
+      endWordIndices.push(i);
+      height = testSpanHeight;
+    }
+  }
+  endWordIndices.push(words.length - 1); // final word index
+  if (moreThanOneLine) {
+    for (let i = 0; i < endWordIndices.length; i += 1) {
+      for (let j = startPoint; j < endWordIndices[i]; j += 1) {
+        line += `${words[j]} `;
+      }
+      lines.push(line);
+      line = '';
+      startPoint = endWordIndices[i];
+    }
+    lines[lines.length - 1] += words[words.length - 1]; // final word
+  }
+  document.body.removeChild(testSpan);
+
+  return lines;
+};
